[Rust] Restructure type APIs into `types` module
This helps with documentation, giving a single module for those working with types to find related APIs

Also split out enumeration and structure APIs into their own file, since they have their own backing data separate from `Type`.
