+++ rust/examples/dump_type_library.rs
@@ -2,8 +2,8 @@
 
 use binaryninja::binary_view::BinaryView;
 use binaryninja::file_metadata::FileMetadata;
-use binaryninja::type_library::TypeLibrary;
-use binaryninja::type_printer::{CoreTypePrinter, TokenEscapingType};
+use binaryninja::types::library::TypeLibrary;
+use binaryninja::types::printer::{CoreTypePrinter, TokenEscapingType};
 
 fn main() {
     let type_lib_str = std::env::args().nth(1).expect("No type library provided");
