+++ rust/src/lib.rs
@@ -81,11 +81,6 @@ pub mod string;
 pub mod symbol;
 pub mod tags;
 pub mod template_simplifier;
-pub mod type_archive;
-pub mod type_container;
-pub mod type_library;
-pub mod type_parser;
-pub mod type_printer;
 pub mod types;
 pub mod update;
 pub mod variable;
