+++ rust/src/platform.rs
@@ -14,15 +14,15 @@
 
 //! Contains all information related to the execution environment of the binary, mainly the calling conventions used
 
-use crate::type_container::TypeContainer;
-use crate::type_parser::{TypeParserError, TypeParserErrorSeverity, TypeParserResult};
 use crate::{
     architecture::{Architecture, CoreArchitecture},
     calling_convention::CoreCallingConvention,
     rc::*,
     string::*,
-    type_library::TypeLibrary,
-    types::QualifiedNameAndType,
+    types::{
+        QualifiedNameAndType, TypeContainer, TypeLibrary, TypeParserError, TypeParserErrorSeverity,
+        TypeParserResult,
+    },
 };
 use binaryninjacore_sys::*;
 use std::fmt::Debug;
