+++ rust/src/types/container.rs
@@ -12,8 +12,7 @@ use crate::platform::Platform;
 use crate::progress::{NoProgressCallback, ProgressCallback};
 use crate::rc::{Array, Ref};
 use crate::string::{raw_to_string, BnString, IntoCStr};
-use crate::type_parser::{TypeParserError, TypeParserResult};
-use crate::types::{QualifiedName, QualifiedNameAndType, Type};
+use crate::types::{QualifiedName, QualifiedNameAndType, Type, TypeParserError, TypeParserResult};
 use binaryninjacore_sys::*;
 use std::collections::HashMap;
 use std::ffi::{c_char, c_void};
