+++ rust/src/types/printer.rs
@@ -5,8 +5,7 @@ use crate::disassembly::InstructionTextToken;
 use crate::platform::Platform;
 use crate::rc::{Array, CoreArrayProvider, CoreArrayProviderInner, Ref};
 use crate::string::{raw_to_string, BnString, IntoCStr};
-use crate::type_container::TypeContainer;
-use crate::types::{NamedTypeReference, QualifiedName, QualifiedNameAndType, Type};
+use crate::types::{NamedTypeReference, QualifiedName, QualifiedNameAndType, Type, TypeContainer};
 use binaryninjacore_sys::*;
 use std::ffi::{c_char, c_int, c_void};
 use std::ptr::NonNull;
